requirements.txt
kafka-python==2.0.2
mysql-connector-python==9.1.0
